requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
fonttools==4.28.1
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
tweepy==4.4.0
urllib3==1.26.7
